+++ BoDi.Performance.Tests/Benchmarks/ResolveAllFromFactory.cs
@@ -12,10 +12,10 @@ namespace BODi.Performance.Tests.Benchmarks
             return Container14.ResolveAll<IAllRegisteredFromFactory>().ToList();
         }
 
-        [Benchmark(Description = "v1.Next-Flawed")]
-        public object Version_1_Next_Flawed()
+        [Benchmark(Description = "v1.BoDi_Concurrent_Dictionary_And_Lazy")]
+        public object Version_1_BoDi_Concurrent_Dictionary_And_Lazy()
         {
-            return Container1NextFlawed.ResolveAll<IAllRegisteredFromFactory>();
+            return Container1Concurrent_Dictionary_And_Lazy.ResolveAll<IAllRegisteredFromFactory>();
         }
 
         [Benchmark(Description = "Current")]
