+++ BoDi.Performance.Tests/Benchmarks/ResolveAllFromType.cs
@@ -12,10 +12,10 @@ namespace BODi.Performance.Tests.Benchmarks
             return Container14.ResolveAll<IAllRegisteredFromType>().ToList();
         }
 
-        [Benchmark(Description = "v1.Next-Flawed")]
-        public object Version_1_Next_Flawed()
+        [Benchmark(Description = "v1.BoDi_Concurrent_Dictionary_And_Lazy")]
+        public object Version_1_BoDi_Concurrent_Dictionary_And_Lazy()
         {
-            return Container1NextFlawed.ResolveAll<IAllRegisteredFromType>();
+            return Container1Concurrent_Dictionary_And_Lazy.ResolveAll<IAllRegisteredFromType>();
         }
 
         [Benchmark(Description = "Current")]
