+++ BoDi.Performance.Tests/Benchmarks/SingleContainerBenchmarkBase.cs
@@ -12,7 +12,7 @@ namespace BODi.Performance.Tests.Benchmarks
     {
         protected internal IObjectContainer ContainerCurrent;
         protected internal BoDi1_4.IObjectContainer Container14;
-        protected internal BoDi1_Next_Flawed.IObjectContainer Container1NextFlawed;
+        protected internal BoDi_Concurrent_Dictionary_And_Lazy.IObjectContainer Container1Concurrent_Dictionary_And_Lazy;
 
         [GlobalSetup]
         public void Setup()
@@ -30,17 +30,17 @@ namespace BODi.Performance.Tests.Benchmarks
             Container14.RegisterFactoryAs<IAllRegisteredFromFactory>(_ => new AllRegistered4());
 
 
-            Container1NextFlawed = new BoDi1_Next_Flawed.ObjectContainer();
-            Container1NextFlawed.RegisterFactoryAs(_ => new FactoryRegistered());
-            Container1NextFlawed.RegisterTypeAs<TypeRegistered, TypeRegistered>();
-            Container1NextFlawed.RegisterTypeAs<AllRegistered1, IAllRegisteredFromType>();
-            Container1NextFlawed.RegisterTypeAs<AllRegistered2, IAllRegisteredFromType>();
-            Container1NextFlawed.RegisterTypeAs<AllRegistered3, IAllRegisteredFromType>();
-            Container1NextFlawed.RegisterTypeAs<AllRegistered4, IAllRegisteredFromType>();
-            Container1NextFlawed.RegisterFactoryAs<IAllRegisteredFromFactory>(_ => new AllRegistered1());
-            Container1NextFlawed.RegisterFactoryAs<IAllRegisteredFromFactory>(_ => new AllRegistered2());
-            Container1NextFlawed.RegisterFactoryAs<IAllRegisteredFromFactory>(_ => new AllRegistered3());
-            Container1NextFlawed.RegisterFactoryAs<IAllRegisteredFromFactory>(_ => new AllRegistered4());
+            Container1Concurrent_Dictionary_And_Lazy = new BoDi_Concurrent_Dictionary_And_Lazy.ObjectContainer();
+            Container1Concurrent_Dictionary_And_Lazy.RegisterFactoryAs(_ => new FactoryRegistered());
+            Container1Concurrent_Dictionary_And_Lazy.RegisterTypeAs<TypeRegistered, TypeRegistered>();
+            Container1Concurrent_Dictionary_And_Lazy.RegisterTypeAs<AllRegistered1, IAllRegisteredFromType>();
+            Container1Concurrent_Dictionary_And_Lazy.RegisterTypeAs<AllRegistered2, IAllRegisteredFromType>();
+            Container1Concurrent_Dictionary_And_Lazy.RegisterTypeAs<AllRegistered3, IAllRegisteredFromType>();
+            Container1Concurrent_Dictionary_And_Lazy.RegisterTypeAs<AllRegistered4, IAllRegisteredFromType>();
+            Container1Concurrent_Dictionary_And_Lazy.RegisterFactoryAs<IAllRegisteredFromFactory>(_ => new AllRegistered1());
+            Container1Concurrent_Dictionary_And_Lazy.RegisterFactoryAs<IAllRegisteredFromFactory>(_ => new AllRegistered2());
+            Container1Concurrent_Dictionary_And_Lazy.RegisterFactoryAs<IAllRegisteredFromFactory>(_ => new AllRegistered3());
+            Container1Concurrent_Dictionary_And_Lazy.RegisterFactoryAs<IAllRegisteredFromFactory>(_ => new AllRegistered4());
 
             ContainerCurrent = new ObjectContainer();
             ContainerCurrent.RegisterFactoryAs(_ => new FactoryRegistered());
